requirements.txt
# requirements.txt

# Flask and extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Babel==2.0.0
Flask-Session==0.4.0
email-validator==1.3.1
python-dotenv==1.0.0
Werkzeug==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2

# Database
SQLAlchemy==2.0.4

# OpenAI
openai==1.1.1

# Document processing
PyPDF2==3.0.1
EbookLib==0.18
python-docx==0.8.11
beautifulsoup4==4.12.2

# Utility packages
requests==2.28.2
black==23.1.0
flake8==6.0.0
pytest==7.3.1

# Deployment
gunicorn==20.1.0